Do not expose the admin port by default
Exposing ports is only useful for publish all ports (`-P`) which will
choose random ports on the host. Since the administrator interface is
only useful when the host and container port for InfluxDB are the same,
this option can never effectively be used. It's better to just remove
the option to prevent confusion.