Include a self-signed certificate in the influxdb image
Generate a self-signed certificate and include it in InfluxDB by
default. This way, someone could start the HTTPS version of InfluxDB for
testing without having to generate their own certificate and then
mounting a filesystem or creating their own image.

Since HTTPS is disabled by default, we aren't shipping an image that
with an insecure certificate. A person who wants to ship with a
secure certificate is going to need to replace the certificate anyway.

HTTPS can be enabled using environment variables so it can use this
self-signed certificate.