Remove port 8083 from the exposed ports in influxdb alpine
It was never supposed to be exposed, but it got copied from the main
image before that port was removed from the main image during initial
development.

Now that the admin interface is deprecated and isn't on by default,
there's even less of a reason for it to be exposed.