Fix the update script and add multi-architecture support
Rather than try to infer everything, the update script is now driven by
a manifest file that lists the versions, variants, and architectures
that should be used.

The script will then read Dockerfiles and validate the structure of the
project before emitting the library files into official-images.

Architecture support has been added so now we can mark an image as
capable of being used with multiple architectures using the same
Dockerfile. If a different directory is required, add a directory with
the name of the architecture into the folder.