bugfix: increment ping_count on loop
The code previously did not inclement the ping_count variable. This could cause an infinite loop, as the loop conditional depends on this value increasing with each time through. This change adds a statement to increment this value at the end of the loop, and also logs the value of this counter instead of the value it is compared to.