fix: telegraf do not run as root
This runs Telegraf as a non-root user. For the Debian based Dockerfiles,
the deb package will create a telegraf user and group for usage. All
that is needed is to say run as the telegraf user.

For Alpine based images, a telegraf user and group needs to be created
and then run as that telegraf user.

Fixes: #412