fix: do not exit on failure of setcap on Telegraf
The Telegraf container entrypoints attempt to use setcap to add
additional capabilites to the telegraf binary. Users who use aufs or run
a kernel that does not allow the setcap option will fail to start the
telegraf container as-is. This allows users to continue to use Telegraf,
but instead print an error message that setcap failed.

Resolves: #561